Person Manager permissions and menu changes
